+++ readium/lcp/r2-lcp/src/main/java/org/readium/r2/lcp/LcpLicense.kt
@@ -6,7 +6,9 @@
 
 package org.readium.r2.lcp
 
-import kotlinx.coroutines.*
+import kotlinx.coroutines.GlobalScope
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.runBlocking
 import org.joda.time.DateTime
 import org.readium.r2.lcp.license.model.LicenseDocument
 import org.readium.r2.lcp.license.model.StatusDocument
@@ -14,7 +16,7 @@ import org.readium.r2.shared.publication.services.ContentProtectionService
 import org.readium.r2.shared.util.Try
 import timber.log.Timber
 import java.net.URL
-import kotlin.coroutines.resume
+import java.util.*
 
 /**
  * Opened license, used to decipher a protected publication and manage its license.
@@ -55,13 +57,12 @@ interface LcpLicense : ContentProtectionService.UserRights {
     val maxRenewDate: DateTime?
 
     /**
-     * Renews the loan up to a certain date (if possible).
+     * Renews the loan by starting a renew LSD interaction.
      *
-     * @param urlPresenter: Used when the renew requires to present an HTML page to the user.
-     *        A reading app should implement it with a `suspendCoroutine` statement to return the
-     *        control once the user dismissed the browser view.
+     * @param prefersWebPage Indicates whether the loan should be renewed through a web page if
+     *        available, instead of programmatically.
      */
-    suspend fun renewLoan(end: DateTime?, urlPresenter: suspend (URL) -> Unit): Try<Unit, LcpException>
+    suspend fun renewLoan(listener: RenewListener, prefersWebPage: Boolean = false): Try<Date?, LcpException>
 
     /**
      * Can the user return the loaned publication?
@@ -79,6 +80,32 @@ interface LcpLicense : ContentProtectionService.UserRights {
     suspend fun decrypt(data: ByteArray): Try<ByteArray, LcpException>
 
 
+    /**
+     * UX delegate for the loan renew LSD interaction.
+     *
+     * If your application fits Material Design guidelines, take a look at [MaterialRenewListener]
+     * for a default implementation.
+     */
+    interface RenewListener {
+
+        /**
+         * Called when the renew interaction allows to customize the end date programmatically.
+         * You can prompt the user for the number of days to renew, for example.
+         *
+         * The returned date can't exceed [maximumDate].
+         */
+        suspend fun preferredEndDate(maximumDate: Date?): Date?
+
+        /**
+         * Called when the renew interaction uses an HTML web page.
+         *
+         * You should present the URL in a Chrome Custom Tab and terminate the function when the
+         * web page is dismissed by the user.
+         */
+        suspend fun openWebPage(url: URL)
+
+    }
+
     @Deprecated("Use `license.encryption.profile` instead", ReplaceWith("license.encryption.profile"))
     val encryptionProfile: String? get() =
         license.encryption.profile
@@ -89,23 +116,11 @@ interface LcpLicense : ContentProtectionService.UserRights {
             .onFailure { Timber.e(it) }
             .getOrNull()
 
-    @Deprecated("Use `renewLoan()` with coroutines instead", ReplaceWith("renewLoan)"))
-    fun renewLoan(end: DateTime?, present: (URL, dismissed: () -> Unit) -> Unit, completion: (LcpException?) -> Unit) {
-        GlobalScope.launch {
-            suspend fun presentUrl(url: URL) {
-                suspendCancellableCoroutine<Unit> { cont ->
-                    present(url) {
-                        if (cont.isActive) {
-                            cont.resume(Unit)
-                        }
-                    }
-                }
-            }
+    @Deprecated("Use `renewLoan` with `RenewListener` instead", ReplaceWith("renewLoan(LcpLicense.RenewListener)"), level = DeprecationLevel.ERROR)
+    suspend fun renewLoan(end: DateTime?, urlPresenter: suspend (URL) -> Unit): Try<Unit, LcpException> = Try.success(Unit)
 
-            val result = renewLoan(end, ::presentUrl)
-            completion(result.exceptionOrNull())
-        }
-    }
+    @Deprecated("Use `renewLoan` with `RenewListener` instead", ReplaceWith("renewLoan(LcpLicense.RenewListener)"), level = DeprecationLevel.ERROR)
+    fun renewLoan(end: DateTime?, present: (URL, dismissed: () -> Unit) -> Unit, completion: (LcpException?) -> Unit) {}
 
     @Deprecated("Use `returnPublication()` with coroutines instead", ReplaceWith("returnPublication"))
     fun returnPublication(completion: (LcpException?) -> Unit) {
@@ -115,7 +130,6 @@ interface LcpLicense : ContentProtectionService.UserRights {
     }
 
 }
-
 
 @Deprecated("Renamed to `LcpService`", replaceWith = ReplaceWith("LcpService"))
 typealias LCPService = LcpService