+++ readium/lcp/CHANGELOG.md
@@ -6,6 +6,12 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## Changed
+
+* The Renew Loan API got revamped to better support renewal through a web page.
+    * You will need to implement `LcpLicense.RenewListener` to coordinate the UX interaction.
+    * If your application fits Material Design guidelines, take a look at `MaterialRenewListener` for a default implementation.
+
 ## [2.0.0-beta.1]
 
 ## Changed